feat: extract canvas pan zoom hook
